Progress update

+++ src/common/noise.h
@@ -5,18 +5,20 @@
     #define NOISE_TABLES 64
 #endif
 
+#include <inttypes.h>
+
 void initNoiseTable(int);
 double noise2d(int, double, double);
 double nnoise2d(int, double, double);
 double perlin2d(int, double, double, double, int);
 double nperlin2d(int, double, double, double, int);
 double noise1(int, double);
-double pnoise1(int, double, int);
+double pnoise1(int, double, int64_t);
 double noise2(int, double, double);
-double pnoise2(int, double, double, int, int);
+double pnoise2(int, double, double, int64_t, int64_t);
 double noise3(int, double, double, double);
-double pnoise3(int, double, double, double, int, int, int);
+double pnoise3(int, double, double, double, int64_t, int64_t, int64_t);
 double noise4(int, double, double, double, double);
-double pnoise4(int, double, double, double, double, int, int, int, int);
+double pnoise4(int, double, double, double, double, int64_t, int64_t, int64_t, int64_t);
 
 #endif
